Provide a default nginx.conf with explicit number of worker_processes (#5)
The provided default nginx config from the alpine nginx image sets worker_processes to auto which in docker will result in us reading the host cpu count and using that many workers. This is sub-optimal for us in most cases so we'll stick to 1 worker process for now. Config is copied from latest alpine image apart from the worker_processes change.